+++ src/Ydb.Sdk/src/Ado/IDriverFactory.cs
@@ -6,6 +6,10 @@ internal interface IDriverFactory
 {
     Task<IDriver> CreateAsync();
 
+    string Endpoint { get; }
+
+    string Database { get; }
+
     string GrpcConnectionString { get; }
 
     ILoggerFactory LoggerFactory { get; }